requirements.txt
pytorch==1.7.1
torchvision==0.8.2
pillow==8.1.0
tensorboardx==2.4
scikit-learn==0.24.1
numpy==1.19.2
opencv==3.4.2
scipy==1.6.1
imageio==2.9.0